+++ src/content/scripting/js_functions.cc
@@ -213,7 +213,7 @@ duk_ret_t js_addCdsObject(duk_context* ctx)
     return 0;
 }
 
-static duk_ret_t convertCharsetGeneric(duk_context* ctx, charset_convert_t chr)
+static duk_ret_t convertCharsetGeneric(duk_context* ctx, CharsetConversion chr)
 {
     auto self = Script::getContextScript(ctx);
     if (duk_get_top(ctx) != 1)
@@ -238,22 +238,22 @@ static duk_ret_t convertCharsetGeneric(duk_context* ctx, charset_convert_t chr)
 
 duk_ret_t js_f2i(duk_context* ctx)
 {
-    return convertCharsetGeneric(ctx, F2I);
+    return convertCharsetGeneric(ctx, CharsetConversion::F2I);
 }
 
 duk_ret_t js_m2i(duk_context* ctx)
 {
-    return convertCharsetGeneric(ctx, M2I);
+    return convertCharsetGeneric(ctx, CharsetConversion::M2I);
 }
 
 duk_ret_t js_p2i(duk_context* ctx)
 {
-    return convertCharsetGeneric(ctx, P2I);
+    return convertCharsetGeneric(ctx, CharsetConversion::P2I);
 }
 
 duk_ret_t js_j2i(duk_context* ctx)
 {
-    return convertCharsetGeneric(ctx, J2I);
+    return convertCharsetGeneric(ctx, CharsetConversion::J2I);
 }
 
 #endif // HAVE_JS