+++ test/scripting/test_external_m3u_playlist.cc
@@ -24,7 +24,6 @@
 
 #include "cds/cds_objects.h"
 #include "upnp/upnp_common.h"
-#include "util/string_converter.h"
 
 #include "mock/common_script_mock.h"
 #include "mock/duk_helper.h"