+++ test/scripting/test_import_script.cc
@@ -24,7 +24,6 @@
 
 #include "content/onlineservice/online_service.h"
 #include "metadata/metadata_handler.h"
-#include "util/string_converter.h"
 
 #include "mock/common_script_mock.h"
 #include "mock/duk_helper.h"